+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2024-06-26
+
 ### dependabot: \#37 Bump braces from 3.0.2 to 3.0.3
 
 - added `dateIsFirstDayOfMonth`, `dateIsLeapYear`, `dateIsWeekend` and `getNextBusinessDate` utility functions
@@ -56,7 +58,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added `eslint-plugin-jsdoc` to lint jsdoc comments
 - `getEnumNameFromValue`,`getEnumValueFromName`, `getEnumNames`,`getEnumNameValues` functions to respectively get the name of an enum from its value, the value from its name, all the names and all the values
 
-[unreleased]: https://github.com/neolution-ch/javascript-utils/compare/0.5.0...HEAD
+[unreleased]: https://github.com/neolution-ch/javascript-utils/compare/0.6.0...HEAD
+[0.6.0]: https://github.com/neolution-ch/javascript-utils/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/neolution-ch/javascript-utils/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/neolution-ch/javascript-utils/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/neolution-ch/javascript-utils/compare/0.3.0...0.3.1